Prevent 16-bit overflow

+++ src/main/telemetry/frsky.h
@@ -5,5 +5,5 @@
 
 #pragma once
 
-uint32_t frskyGetFlightMode(void);
+uint16_t frskyGetFlightMode(void);
 uint16_t frskyGetGPSState(void);
